refactor: 내 정보 페이지 컴포넌트 분리
  - 내 정보와 회원 탈퇴 섹션 분리
