+++ frontend/src/components/MyInformation/index.ts
@@ -1,0 +1,2 @@
+export { default as Info } from './Info';
+export { default as Withdrawal } from './Withdrawal';
